Moved some files to other packages
