+++ core/src/main/scala/it/parttimeteam/core/GameState.scala
@@ -1,0 +1,3 @@
+package it.parttimeteam.core class GameState {
+
+}
