+++ core/src/main/scala/it/parttimeteam/core/collections/Board.scala
